+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/Style.java
@@ -518,17 +518,17 @@ public class Style {
    */
   void clear() {
     fullyLoaded = false;
-    for (Source source : sources.values()) {
-      if (source != null) {
-        source.setDetached();
-        nativeMap.removeSource(source);
-      }
-    }
-
     for (Layer layer : layers.values()) {
       if (layer != null) {
         layer.setDetached();
         nativeMap.removeLayer(layer);
+      }
+    }
+
+    for (Source source : sources.values()) {
+      if (source != null) {
+        source.setDetached();
+        nativeMap.removeSource(source);
       }
     }